Updates!
Added a mousePressed and mouseReleased function for interactivity with the Box. (Hoping to make it more interactive!)

When mouse is pressed, the boxes spread out over a larger area.
When mouse is released, the boxes return to their original state.

I also added a spotLight, its angle controlled by the mouseX and mouseY position. The lights randomize every loop, giving it a very hypnotic effect.